pyproject.toml
[project]
name = "recallify"
version = "0.1.0"
description = "Recallify application"
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.10.4",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.0",
    "boto3>=1.34.0",
    "pygithub>=2.1.1",
    "slack-sdk>=3.26.0",
    "httpx>=0.26.0",
    "xid>=1.0.0",
    "click>=8.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
